For FutureLearn Creative Coding
My aim with this was to make an sort of generative ransom note.
When you run the sketch, you get a blank screen. If you hit letter keys, letters appear on the screen, although not the ones you press! The letters are generated randomly with the same frequency as a Scrabble set, and they wander round the screen, generally settling in two lines and trying to form sequences. Other key strokes have some effect as well.
The fonts were chosen to look like they were cut out from newspaper headlines (or like the cover of the Sex Pistols first album).